refactor: move resources::usage types from lib to wire crate
