+++ lib/src/resources/usage.rs
@@ -1,55 +1,9 @@
 use crate::common::{request, SerializableNone};
 use crate::error::ApiError;
+use lrzcc_wire::resources::CloudUsage;
 use reqwest::blocking::Client;
 use reqwest::{Method, StatusCode};
-use serde::{Deserialize, Serialize};
 use std::rc::Rc;
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct CloudUsageOverviewInner {
-    pub total: u64,
-    pub used: u64,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct CloudUsageOverviewInnerFloat {
-    pub total: f64,
-    pub used: f64,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct CloudUsageOverview {
-    pub vcpus: CloudUsageOverviewInner,
-    pub ram: CloudUsageOverviewInner,
-    pub gpus: CloudUsageOverviewInner,
-    pub storage: CloudUsageOverviewInnerFloat,
-    pub mwn_ips: CloudUsageOverviewInner,
-    pub www_ips: CloudUsageOverviewInner,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct CloudUsageFlavorSlot {
-    pub name: String,
-    pub free: u32,
-    pub total: u32,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct CloudUsageAggregate {
-    pub name: String,
-    pub title: String,
-    pub flavors: Vec<CloudUsageFlavorSlot>,
-}
-
-// TODO how could we handle a table representation of this?
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct CloudUsage {
-    pub overview: CloudUsageOverview,
-    pub lrz_flavor_slots: Vec<CloudUsageAggregate>,
-    pub ach_flavor_slots: Vec<CloudUsageAggregate>,
-    pub other_flavor_slots: Vec<CloudUsageAggregate>,
-    pub datetime: String,
-}
 
 pub struct UsageApi {
     pub url: String,
